requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.4.26
click==8.1.8
colorama==0.4.6
docutils==0.21.2
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.16.0
idna==3.10
lxml==5.4.0
natsort==8.4.0
outcome==1.3.0.post0
overrides==7.7.0
protobuf==5.29.3
PySocks==1.7.1
robotframework==7.2.2
robotframework-assertion-engine==3.0.3
robotframework-browser==19.4.0
robotframework-pandaslibrary==0.0.1
robotframework-pythonlibcore==4.4.1
robotframework-seleniumlibrary==6.7.1
rpa==1.50.0
seedir==0.5.0
selenium==4.31.0
setuptools==80.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
tagui==1.50.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
